requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
cx-Freeze==6.11.1
httptools==0.5.0
importlib-metadata==3.10.0
lief==0.12.2
multidict==5.1.0
packaging==21.3
Pillow==9.2.0
pycparser==2.20
pyparsing==3.0.9
pystray==0.17.3
python-osc==1.8.0
pywin32==304
sanic==22.6.2
sanic-routing==22.3.0
six==1.15.0
typing-extensions==3.7.4.3
websockets==10.3
zipp==3.4.1
